requirements.txt
openai-whisper
ebmlite
requests
python-dotenv